fixed responsive issue, header content issue and some other things mentioned in sheets
