+++ .env
@@ -11,3 +11,4 @@ LINKEDIN_URL=https://www.linkedin.com/company/gitlime/
 MAIL=hello@gitlime.com
 PINTREST_URL=https://in.pinterest.com/gitlime
 STRAPI_API_URL=https://content.abhishek.world/
+ORIGIN_URL=http://localhost:3001
